Reservation is done !!

+++ src/model/User.java
@@ -7,7 +7,7 @@ public class User {
 	private String _firstName;
 	private String _lastName;
 	private String _type;
-	private int _balance;
+	private long _balance;
 	private String _company;
 	
 	public User(String username, String email, String password, String firstName, 
@@ -58,10 +58,10 @@ public class User {
 	public void setType(String type) {
 		this._type = type;
 	}
-	public int getBalance() {
+	public long getBalance() {
 		return _balance;
 	}
-	public void setBalance(int balance) {
+	public void setBalance(long balance) {
 		this._balance = balance;
 	}
 	public String getCompnay() {
